pyproject.toml
[tool.poetry]
name = "feedgy-ds-repo-template"
version = "0.0.1"
description = "Data science git repo template"
authors = ["Wenkun LI <wli@feedgy.solar>"]
readme = "README.md"
packages = [{ include = "feedgy_ds_repo_template", from = "src" }]


[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.76"
ipykernel = "^6.29.4"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
ruff = "^0.3.4"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = 'feedgy_ds_repo_template.main:run'

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = [
  "E501",  # line-too-long
  "UP",    # pyupgrade
  "D",     # pydocstyle
]
ignore = [
  "D100", # Missing docstring in public module
  "D103", # Missing docstring in public function
  "D104", # Missing docstring in public package
]

[tool.ruff.lint.pydocstyle]
convention = "google"
